requirements.txt
pandas~=2.0.2
python-dotenv~=1.0.0